backend_code/requirements.txt
absl-py==2.1.0
alembic==1.14.1
astunparse==1.6.3
beautifulsoup4==4.13.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dlib==19.24.6
dnspython==2.7.0
filelock==3.17.0
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.1.24
gast==0.6.0
gdown==5.2.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h5py==3.12.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
proto-plus==1.26.0
protobuf==3.20.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.1
PyMySQL==1.1.1
pyparsing==3.2.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.37
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.22.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tf-estimator-nightly==2.8.0.dev2021122109
tqdm==4.67.1
typeguard==2.13.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
